% Generated by roxygen2 (4.0.2): do not edit by hand
\name{glmm}
\alias{glmm}
\title{Fit Generalized Linear Mixed-Effects Models Using State Space Framework}
\usage{
glmm(response.var, grouping.var, fixed.formula, random.formula, data,
  distribution, init.random, init.dispersion, init.fixed,
  correlating.effects = TRUE, REML = TRUE, estimate.dispersion,
  return.model = TRUE, nsim = 0, maxiter = 50, maxeval = 1000,
  xtol_rel = 1e-06, ...)
}
\arguments{
\item{response.var}{Name of the response variable in data.}

\item{grouping.var}{Name of the grouping variable in data. Only one grouping variable is allowed.}

\item{data}{Optional data frame environment containing the variables in the model.}

\item{distribution}{Character vector defining the distributions used for each group.
Either length of one (same distribution for all groups) or length of p, number of groups. Default is "gaussian".}

\item{init.random}{Initial values for random effect covariances.}

\item{init.dispersion}{Initial values for dispersion paremeters.}

\item{init.fixed}{Initial values for fixed effects.}

\item{correlating.effects}{Logical. Default is TRUE.}

\item{estimate.dispersion}{Logical. Is the dispersion parameter estimated or fixed.}

\item{return.model}{Logical. Is the estimated model returned in output.}

\item{nsim}{Integer. Number of independent samples used in importance sampling.
Default is 0, which corresponds to Laplace approximation.}

\item{maxiter}{Integer. Number of iterations for in iterative weighted least squares.}

\item{method}{"REML", "REML-ML" or "ML". "REML-ML" uses estimates obtained from REML for ML estimation.
Note that ML is much slower.}
}
\description{
Function \code{glmm} estimates GLMM using methods basec on state space modelling.
}
\examples{
data(butterfly)

require(lme4)
system.time(fit.SSglmm.reml<-glmm(response.var="Colias", grouping.var="site", fixed=~habitat + building + urbanveg,random=~1,
                                 REML=TRUE,data=butterfly,distribution="poisson",return.model=TRUE))

system.time(fit.SSglmm.ml<-glmm(response.var="Colias", grouping.var="site", fixed=~habitat + building + urbanveg,random=~1,
                               REML=FALSE,data=butterfly,distribution="poisson",return.model=TRUE))

fit.glmer<-glmer(Colias ~ habitat + building + urbanveg + (1|site),
                family=poisson, data=butterfly,control=glmerControl(optimizer = "bobyqa"))

sqrt(fit.SSglmm.reml$random$P)
sqrt(fit.SSglmm.ml$random$P)
VarCorr(fit.glmer)

fixef(fit.glmer)
fit.SSglmm.ml$fixed$coef
fit.SSglmm.reml$fixed$coef

require(mvabund)
data(spider)
x<-apply(spider$x,2,rep,times=12)
dataspider<-data.frame(species=rep(names(spider$abund),each=28),abund=unlist(spider$abund),x)
out<-glmm(response.var="abund", grouping.var="species", fixed=~soil.dry+bare.sand, random=~1,
data=dataspider,distribution="poisson")

out2<-glmm(response.var="abund", grouping.var="species", fixed=~soil.dry+bare.sand,
data=dataspider,distribution="poisson")
fit<-manyglm(mvabund(spider$abund)~soil.dry+bare.sand,data=data.frame(spider$x),family="poisson")
}

